separate manifest controlled toolbar styling to abstract class
this will allow articles to include some navbar styling in the manifest
